+++ utils/column_names/ind_cqc_pipeline_columns.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from typing import Final
 
 from utils.column_names.capacity_tracker_columns import (
     CapacityTrackerCareHomeCleanColumns as CTCHClean,
@@ -187,6 +188,7 @@ class IndCqcColumns:
     estimate_source: str = "estimate_source"
     estimate_value: str = "estimate_value"
     expected_filled_posts: str = "expected_filled_posts"
+    EXPANDED_ID: Final[str] = "id_per_locationid_import_date_job_role"
     extrapolation_backwards: str = "extrapolation_backwards"
     extrapolation_forwards: str = "extrapolation_forwards"
     extrapolation_model: str = "extrapolation_model"
@@ -311,6 +313,7 @@ class IndCqcColumns:
         "residuals_estimate_filled_posts_non_res_pir"
     )
     rmse: str = "rmse"
+    ROW_ID: Final[str] = "id_per_locationid_import_date"
     service_count: str = "service_count"
     service_count_capped: str = "service_count_capped"
     services_offered: str = CQCLClean.services_offered